BL-635 merged three data reducers into one, adjusted all state-mapping functions accordingly, fixed another missing key error from a map function, removed the obviated apiDataReducer file, restructured the code for consistency.
