+++ src/state/reducers/dataReducer.js
@@ -1,34 +1,50 @@
-import { combineReducers } from 'redux';
-import { GET_DATA, FILTER_SEARCH, RESET_DATA } from '../actionTypes';
+import {
+  FILTER_SEARCH,
+  RESET_DATA,
+  GET_DEFAULT_COMPARISONS,
+  GET_FILTERED_DATA,
+  GET_MOCK_FILTERED_DATA,
+} from '../actionTypes';
 
-// States
-const initialCasesData = [];
-
-const initialFilteredStates = {
-  count: 0,
-  data: [],
+const initialState = {
+  cases: [],
+  comparisonData: {
+    //This will hold keys and values for default displays on the site landing page, similar to TRAC
+  },
+  filterCount: 0,
+  filteredCases: [],
 };
-
-// Reducers
-function casesReducer(state = initialCasesData, action) {
+function dataReducer(state = initialState, action) {
   switch (action.type) {
-    case GET_DATA: {
-      return action.payload;
+    case GET_DEFAULT_COMPARISONS: {
+      //When memoized comparison data is fetched from the server to display on the site landing page
+      return {
+        ...state,
+        comparisonData: action.payload,
+      };
     }
-
-    default: {
-      return state;
+    //When a filter querystring is passed to the server and case data is returned
+    case GET_FILTERED_DATA: {
+      return {
+        ...state,
+        cases: action.payload,
+      };
     }
-  }
-}
-
-function filteredReducer(state = initialFilteredStates, action) {
-  switch (action.type) {
+    case GET_MOCK_FILTERED_DATA: {
+      return {
+        ...state,
+        cases: action.payload,
+      };
+    }
     case FILTER_SEARCH: {
-      return { ...state, data: action.payload, count: state.count + 1 };
+      return {
+        ...state,
+        filteredCases: action.payload,
+        filterCount: state.count + 1,
+      };
     }
     case RESET_DATA: {
-      return initialFilteredStates;
+      return { ...state, filterCount: 0, filteredCases: [] };
     }
     default: {
       return state;
@@ -36,4 +52,4 @@ function filteredReducer(state = initialFilteredStates, action) {
   }
 }
 
-export default combineReducers({ casesReducer, filteredReducer });
+export default dataReducer;
